fix: Potential Vulnerabilities Identified Through Static Code Analysis (#1736)
* removed rand.Seed usage

* removed few type conflicts + returned back linter rule G115

* few more gosec G115 fixes

* few fixes and comments added

* minor changes

* added //nolint:gosec in time conversions which can't be resolved

* added nolint:gosec reason

* updated to golangcli-lint 1.61.0

* super minor refactoring uint64 for cfg.GetValidatorStats()

* refactoring

* added conversion wrapper to devide all existing G115 issues

* refactored many G115 issues

* minor fix

* fixed expected value returned from mock

* removed unused code

* renames

* cr: role to int32

* cr: remove len conversion, add duration conversion check, create var for cutoffround

* fix

* rename package from conversion to casts

* readability

* remove unnecessary loop gymnastics

* remove more loop gymnastics

* add nosec for trusted subnet input

* approve spec diffs

* approve spec diff

* approve genesis spec diff

* fix: duty count uint64

* fix test

* remove extra debug log

---------

+++ network/discovery/dv5_service.go
@@ -209,7 +209,7 @@ func (dvs *DiscV5Service) initDiscV5Listener(logger *zap.Logger, discOpts *Optio
 	dvs.bootnodes = dv5Cfg.Bootnodes
 
 	logger.Debug("started discv5 listener (UDP)", fields.BindIP(bindIP),
-		zap.Int("UdpPort", opts.Port), fields.ENRLocalNode(localNode), fields.Domain(discOpts.DomainType.DomainType()))
+		zap.Uint16("UdpPort", opts.Port), fields.ENRLocalNode(localNode), fields.Domain(discOpts.DomainType.DomainType()))
 
 	return nil
 }
@@ -260,7 +260,7 @@ func (dvs *DiscV5Service) discover(ctx context.Context, handler HandleNewPeer, i
 }
 
 // RegisterSubnets adds the given subnets and publish the updated node record
-func (dvs *DiscV5Service) RegisterSubnets(logger *zap.Logger, subnets ...int) (updated bool, err error) {
+func (dvs *DiscV5Service) RegisterSubnets(logger *zap.Logger, subnets ...uint64) (updated bool, err error) {
 	if len(subnets) == 0 {
 		return false, nil
 	}
@@ -277,7 +277,7 @@ func (dvs *DiscV5Service) RegisterSubnets(logger *zap.Logger, subnets ...int) (u
 }
 
 // DeregisterSubnets removes the given subnets and publish the updated node record
-func (dvs *DiscV5Service) DeregisterSubnets(logger *zap.Logger, subnets ...int) (updated bool, err error) {
+func (dvs *DiscV5Service) DeregisterSubnets(logger *zap.Logger, subnets ...uint64) (updated bool, err error) {
 	logger = logger.Named(logging.NameDiscoveryService)
 
 	if len(subnets) == 0 {
@@ -384,10 +384,10 @@ func (dvs *DiscV5Service) createLocalNode(logger *zap.Logger, discOpts *Options,
 }
 
 // newUDPListener creates a udp server
-func newUDPListener(bindIP net.IP, port int, network string) (*net.UDPConn, error) {
+func newUDPListener(bindIP net.IP, port uint16, network string) (*net.UDPConn, error) {
 	udpAddr := &net.UDPAddr{
 		IP:   bindIP,
-		Port: port,
+		Port: int(port),
 	}
 	conn, err := net.ListenUDP(network, udpAddr)
 	if err != nil {
