+++ network/discovery/options.go
@@ -28,9 +28,9 @@ type DiscV5Options struct {
 	// BindIP is the IP to bind to the UDP listener
 	BindIP string
 	// Port is the UDP port used by discv5
-	Port int
+	Port uint16
 	// TCPPort is the TCP port exposed in the ENR
-	TCPPort int
+	TCPPort uint16
 	// NetworkKey is the private key used to create the peer.ID if the node
 	NetworkKey *ecdsa.PrivateKey
 	// Bootnodes is a list of bootstrapper nodes
